refactor: remove eslint ignore
