+++ tasklist/client/src/pages/Layout/Header/HeaderV2/useSidebarChildren.tsx
@@ -26,7 +26,7 @@ import {getStateLocally} from 'modules/local-storage';
 import {isForbidden} from 'modules/utils/isForbidden';
 import {useMemo} from 'react';
 import {matchPath, useLocation, useSearchParams} from 'react-router-dom';
-import {useCustomFiltersContext} from 'modules/tasks/available-tasks/CollapsiblePanel/CustomFiltersModal/CustomFiltersProvider';
+import {useCustomFiltersContext} from 'modules/tasks/available-tasks/CollapsiblePanel/CustomFiltersModal/CustomFiltersContext';
 import {IconButton} from '@carbon/react';
 import styles from '../styles.module.scss';
 import {getNavLinkSearchParam} from 'modules/features/tasks/filters/getNavLinkSearchParam';
